style: rename components using PascalCase
Just for consistency going forward - per Airbnb style.
