+++ client/components/index.js
@@ -3,8 +3,8 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
+export {default as Navbar} from './NavBar'
+export {default as UserHome} from './UserHome'
+export {Login, Signup} from './AuthForm'
 export {default as LandingPage} from './LandingPage'
 export {default as NoneFound} from './NoneFound'
